backend/requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
fsspec==2025.3.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
opencv-contrib-python==4.11.0.86
pillow==11.1.0
requests==2.32.3
segment_anything @ git+https://github.com/facebookresearch/segment-anything.git@dca509fe793f601edb92606367a655c15ac00fdf
setuptools==78.1.0
sympy==1.13.1
torch==2.6.0
torchvision==0.21.0
typing_extensions==4.13.0
urllib3==2.3.0
Werkzeug==3.1.3
